Upload All

+++ FlexerApp/Models/LoginModel.cs
@@ -10,11 +10,11 @@ namespace FlexerApp.Models
         public string LocationType { get; set; }
         public string IPAddress { get; set; }
         public string City { get; set; }
-        public float Lat { get; set; }
-        public float Long { get; set; }
+        public decimal Lat { get; set; }
+        public decimal Long { get; set; }
         public int sessionID { get; set; }
         public string loginToken { get; set; }
-        public float GMTDiff { get; set; }
+        public decimal GMTDiff { get; set; }
         public DateTime LoginDate { get; set; }
         public DateTime CreatedDate { get; set; }
     }
